+++ test/transfer.ts
@@ -4,7 +4,7 @@ const Scalar = require('ffjavascript').Scalar;
 import { Account } from '../helper.ts/account';
 import { hashAccountState } from '../helper.ts/state-utils';
 import { SimpleTest, TestComponent } from './interface';
-import { TxType } from './common';
+import { TxType, getBTreeProof } from './common';
 
 // circuit-level definitions
 const balanceLevels = 5;
@@ -128,34 +128,6 @@ function initTestCase() {
       tmpAccountRoot: tmpAccountProof.root,
       newAccountRoot: newAccountProof.root,
     };
-}
-
-function getBTreeProof(leaves, index) {
-  // TODO: assert even length
-  // TODO: check index bounds
-
-  let tmpLeaves = leaves;
-  let path_elements = [];
-
-  while(tmpLeaves.length != 1){
-    if (index%2==0){
-      path_elements.push([tmpLeaves[index+1]]);
-    } else {
-      path_elements.push([tmpLeaves[index-1]]);
-    };
-
-    let tempMidLeaves = [];
-    for (let i = 0; (i+1) < tmpLeaves.length; i+=2) {
-      tempMidLeaves.push(hash([tmpLeaves[i], tmpLeaves[i+1]]));
-    }
-    tmpLeaves = tempMidLeaves;
-    index = Math.trunc(index/2);
-  }
-
-  return {
-    root: tmpLeaves[0],
-    path_elements: path_elements,
-  }
 }
 
 let test_case = initTestCase();
